refactor(security): 重构刷新令牌流程
- 移除 OnlineLoginUser 中的 IP 字段，改为 ip
- 优化 refreshToken 方法，简化流程并提高安全性
- 更新相关控制器和工具类以适应新的刷新流程
- 调整 ResponseUtils 类，增加 status 参数

+++ chuang-common/chuang-common-redis/src/main/java/cn/zhangchuangla/common/redis/core/RedisCache.java
@@ -61,7 +61,7 @@ public class RedisCache {
      * @param value   value
      * @param timeout 超时时间（秒）
      */
-    public <T> void setCacheObject(final String key, final T value, final int timeout) {
+    public <T> void setCacheObject(final String key, final T value, final long timeout) {
         redisTemplate.opsForValue().set(key, value, timeout, TimeUnit.SECONDS);
     }
